requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
python-dotenv
dvc
mlflow
bentoml
flask
smote
imbalanced-learn
bentoml
fastapi
uvicorn
pandas
numpy
pyyaml
-e .
